Expand getFarcasterUsers lib to support wallets and fid
